move reaction handling for help to base cog
