+++ cogs/base.py
@@ -69,7 +69,7 @@ class Base(commands.Cog):
     @commands.cooldown(rate=2, per=60.0, type=commands.BucketType.user)
     @commands.command(aliases=["help"])
     async def god(self, ctx):
-        embed = self.reaction.make_embed(1)
+        embed = self.make_embed(1)
 
         channel = await self.check.get_room(ctx.message)
         if channel is not None and channel.id != config.bot_room:
@@ -84,6 +84,57 @@ class Base(commands.Cog):
         if len(messages.info) > 1:
             await msg.add_reaction("◀")
             await msg.add_reaction("▶")
+
+    @commands.Cog.listener()
+    async def on_raw_reaction_add(self, payload):
+        ctx = await utils.reaction_get_ctx(self.bot, payload)
+        if ctx is None:
+            return
+
+        if ctx['message'].embeds and ctx['message'].embeds[0].title == "Rubbergod":
+            if ctx['emoji'] in ["◀", "▶"]:
+                page = int(ctx['message'].embeds[0].footer.text[5])
+                next_page = utils.pagination_next(ctx['emoji'], page, len(messages.info))
+                if next_page:
+                    embed = self.make_embed(next_page)
+                    await ctx['message'].edit(embed=embed)
+            try:
+                await ctx['message'].remove_reaction(ctx['emoji'], ctx['member'])
+            except Exception:
+                pass
+
+    def make_embed(self, page):
+        embed = discord.Embed(title="Rubbergod",
+                              description="Nejlepší a nejúžasnější bot ever.",
+                              color=0xeee657)
+
+        prefix = config.default_prefix
+
+        embed.add_field(name="Autor", value="Toaster#1111")
+
+        # Shows the number of servers the bot is member of.
+        embed.add_field(name="Počet serverů s touto instancí bota",
+                        value=f"{len(self.bot.guilds)}")
+
+        embed.add_field(name="\u200b", value="Příkazy:", inline=False)
+
+        info = messages.info[page - 1]
+
+        for command in info:
+            embed.add_field(name=prefix + command[0],
+                            value=command[1],
+                            inline=False)
+
+        git_hash = utils.git_hash()
+        footer_text = f"Commit {git_hash}"
+
+        if len(messages.info) > 1:
+            footer_text = f"Page {page} | {footer_text}"
+
+        embed.set_footer(text=footer_text, icon_url="https://cdn.discordapp.com/avatars/"
+                         "560917571663298568/b93e8c1e93c2d18b"
+                         "fbd226a0b614cf57.png?size=32")
+        return embed
 
 
 def setup(bot):
