+++ cogs/karma.py
@@ -22,17 +22,6 @@ class Karma(commands.Cog):
         self.reaction = reaction.Reaction(bot, karma_r)
 
     @commands.Cog.listener()
-    async def on_message(self, message):
-
-        if message.author.bot:
-            return
-
-        if message.content.startswith(config.role_string) or\
-           message.channel.id in config.role_channels:
-            role_data = await self.reaction.get_join_role_data(message)
-            await self.reaction.message_role_reactions(message, role_data)
-
-    @commands.Cog.listener()
     async def on_raw_reaction_add(self, payload):
         try:
             await self.reaction.add(payload)
